# Specialized set_dot for one-entry vector