--- a/src/dual_equality_constraints.jl
+++ b/src/dual_equality_constraints.jl
@@ -409,9 +409,6 @@ struct OneEntryVector{T} <: AbstractVector{T}
     value::T
     index::Int
     n::Int
-    function OneEntryVector{T}(value::T, index::Integer, n::Integer)
-        return new{T}(value, index, n)
-    end
 end
 Base.eltype(::Type{OneEntryVector{T}}) where {T} = T
 Base.length(v::OneEntryVector) = v.n
@@ -442,7 +439,7 @@ function MOI.Utilities.triangle_dot(
 end
 
 function set_dot(i::Integer, s::MOI.AbstractVectorSet, T::Type)
-    vec = OneEntryVector(one(T), i, MOI.dimension(s))
+    vec = OneEntryVector{T}(one(T), i, MOI.dimension(s))
     return MOIU.set_dot(vec, vec, s)
 end
 function set_dot(::Integer, ::MOI.AbstractScalarSet, T::Type)