On the benchmark of https://github.com/jump-dev/MathOptInterface.jl/issues/1796, in terms of allocations:
|                | theta1    | thetaG11   |
|----------------|-----------|------------|
| Before this PR | 14.693 MB | 793.687 GB |
| After this PR  | 0.395 MB | 0.039 GB   |

and in terms of time:

|                | theta1    | thetaG11   |
|----------------|-----------|------------|
| Before this PR | 28.5 ms | 232 s |
| After this PR  | 0.586 ms | 0.108 s   |

So for thetaG11, this gives a 20000x speedup in allocation and 2000x speedup in time.

Closes https://github.com/jump-dev/Dualization.jl/issues/66