@guilhermebodin 2022-04-01
Solves https://github.com/jump-dev/Dualization.jl/issues/66